refactor: use typechain solely
